requirements.txt
pdfplumber
pandas
openpyxl
pytesseract
pypdfium2
Pillow
camelot-py
opencv-python-headless
img2table
python-dotenv
requests

# DSPy Framework and OOP Infrastructure
dspy-ai
pydantic
pydantic-settings
click